Initial commit
Run once solution for notification of streams going down. Uses the TwitchAPI to determine if a stream is active or not. Will send an iMessage to all contacts every 5 minutes once a stream is detected as down until the stream is either brought back online or the script is terminated.